Make perf map symbolization test an end-to-end test
Make the perf map symbolization test an end-to-end test and move it into
the corresponding test suite. Instead of exposing our PerfMap logic,
implement a ghetto perf map line parsing logic in five or so lines
inside the test itself.
